pyproject.toml
[project]
name = "car-tracker"
version = "0.1.0"
description = "Costco Travel car rental price scraper"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "jinja2>=3.1.6",
    "nodriver>=0.48.1",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "uvicorn[standard]>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/car_tracker"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.6",
]
